Change type of comparison counters from float to int

+++ src/main/java/org/scify/jedai/blockprocessing/AbstractBlockProcessing.java
@@ -30,7 +30,7 @@ import java.util.List;
 public abstract class AbstractBlockProcessing implements IBlockProcessing {
     
     protected void printOriginalStatistics(List<AbstractBlock> inputBlocks) {
-        float comparisons = 0;
+        int comparisons = 0;
         comparisons = inputBlocks.stream().map((block) -> block.getNoOfComparisons()).reduce(comparisons, (accumulator, _item) -> accumulator + _item);
         
         Log.info("Original blocks\t:\t" + inputBlocks.size());
