+++ src/main/java/org/scify/jedai/datamodel/BilateralBlock.java
@@ -39,7 +39,7 @@ public class BilateralBlock extends AbstractBlock implements Serializable {
         super(entropy);
         index1Entities = entities1;
         index2Entities = entities2;
-        comparisons = ((float) index1Entities.length) * ((float) index2Entities.length);
+        comparisons = index1Entities.length * index2Entities.length;
     }
 
     @Override
