+++ src/test/java/org/scify/jedai/configuration/StepByStepRandomConfigurationDER.java
@@ -50,8 +50,8 @@ public class StepByStepRandomConfigurationDER {
     private final static int NO_OF_TRIALS = 100;
 
     static float getTotalComparisons(List<AbstractBlock> blocks) {
-        float originalComparisons = 0;
-        originalComparisons = blocks.stream().map(AbstractBlock::getNoOfComparisons).reduce(originalComparisons, Float::sum);
+        int originalComparisons = 0;
+        originalComparisons = blocks.stream().map(AbstractBlock::getNoOfComparisons).reduce(originalComparisons, Integer::sum);
         System.out.println("Original comparisons\t:\t" + originalComparisons);
         return originalComparisons;
     }
